Change case limit type from int to size_t.

+++ src/data/dictionary.h
@@ -74,8 +74,8 @@ void dict_set_weight (struct dictionary *, struct variable *);
 struct variable *dict_get_filter (const struct dictionary *);
 void dict_set_filter (struct dictionary *, struct variable *);
 
-int dict_get_case_limit (const struct dictionary *);
-void dict_set_case_limit (struct dictionary *, int);
+size_t dict_get_case_limit (const struct dictionary *);
+void dict_set_case_limit (struct dictionary *, size_t);
 
 int dict_get_next_value_idx (const struct dictionary *);
 size_t dict_get_case_size (const struct dictionary *);
